[PATCH] add some feature for H5 build scripts

+++ lib/H5SDK_BuildEnvironment.sh
@@ -89,14 +89,21 @@ function git_configure()
 
 function install_toolchain()
 {
-    mkdir -p $TOP_ROOT/OrangePiH5/.tmp_toolchain
-    cd $TOP_ROOT/OrangePiH5/.tmp_toolchain
-    curl -C - -o ./toolchain $toolchain
-    unzip $TOP_ROOT/OrangePiH5/.tmp_toolchain/toolchain
-    mkdir -p $TOP_ROOT/OrangePiH5/toolchain
-    mv $TOP_ROOT/OrangePiH5/.tmp_toolchain/OrangePiH5_toolchain-master $TOP_ROOT/OrangePiH5/toolchain/toolchain_tar
-    rm -rf $TOP_ROOT/OrangePiH5/.tmp_toolchain
-    cd -
+    if [ ! -d $TOP_ROOT/OrangePiH5/toolchain/gcc-linaro-aarch ]; then
+        mkdir -p $TOP_ROOT/OrangePiH5/.tmp_toolchain
+        cd $TOP_ROOT/OrangePiH5/.tmp_toolchain
+        for ((i=0; i < 100; i++)) do
+            curl -C - -o ./toolchain $toolchain
+            if [ -f $TOP_ROOT/OrangePiA64/.tmp_toolchain/toolchain ]; then
+                i=200
+            fi
+        done
+        unzip $TOP_ROOT/OrangePiH5/.tmp_toolchain/toolchain
+        mkdir -p $TOP_ROOT/OrangePiH5/toolchain
+        mv $TOP_ROOT/OrangePiH5/.tmp_toolchain/OrangePiH5_toolchain-master $TOP_ROOT/OrangePiH5/toolchain/toolchain_tar
+        rm -rf $TOP_ROOT/OrangePiH5/.tmp_toolchain
+        cd -
+    fi
 }
 
 git_configure
@@ -105,4 +112,7 @@ dirent_check
 install_toolchain
 end_op
 
-cd $TOP_ROOT
+whiptail --title "OrangePi Build System" --msgbox \
+ "`figlet OrangePi` Succeed to Create OrangePi Build System!        Path:$TOP_ROOT/OrangePiH5" \
+            15 50 0
+clear
